Member sync improvements - Refactors logic to command instead of a static class - Move sync table lookups to a single call instead of iteratively - Fix dateetime discrepancies with epoch - Cache division ids as they are less likely to change
